requirements.txt
# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Google Sheets integration
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Web scraping with Selenium
selenium>=4.0.0
webdriver-manager>=3.8.0

# Telegram bot
python-telegram-bot>=20.0
python-telegram-bot[job-queue]

# Standard library modules (no installation needed)
# - logging (built-in)
# - datetime (built-in)
# - random (built-in)
# - time (built-in)
# - os (built-in)
# - re (built-in)
# - urllib.parse (built-in)